requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
pika==1.3.2
openai==1.3.8
loguru==0.7.2
prometheus-client==0.19.0
structlog==23.2.0
bcrypt==4.1.2
email-validator==2.1.0
jinja2==3.1.2